Fix to match Aura.Di 4.x
Although `composer.json` specifies a `require-dev` which includes this line:
```
        "aura/di": "~4.0",
```
Aura.Auth does not work properly with Aura.Di 4.x.  This is because strict type checking in PHP 7.2+ requires that the function signatures for `define()` and `modify()` specify a `void` return type.  The PHPUnit tests apparently do not instantiate any object using this provided configuration file (`Common.php`), because to do so throws a fatal error.  This change should fix that problem.